Rename intro directory (switch ":" for "-" for Windows compatibility)
User humza in Slack reported a clone error on Windows:

`fatal: cannot create directory at '01 - Introduction : Start Here': Invalid argument`

This PR renames the introduction to remove the colon, which is a
reserved file name character in Windows. I do not have a running
Windows install to test this on but it should theoretically fix the
problem.